♻️(recruteur-infrastructure) split models
