+++ src/web/infrastructure/django_apps/recruteur/models/organisme.py
@@ -1,0 +1,31 @@
+from django.db import models
+from referentiel.value_objects.verse import Verse
+
+from infrastructure.django_apps.utils.models import BaseDatedModel
+
+
+class OrganismeModel(BaseDatedModel):
+    nom = models.CharField(max_length=255)
+    versant = models.CharField(
+        max_length=10,
+        choices=[(v.value, v.value) for v in Verse],
+    )
+    siret = models.CharField(max_length=14, unique=True)
+    parent_id = models.UUIDField(null=True, blank=True)
+    localisation = models.JSONField(null=True, blank=True)
+    etapes = models.JSONField(
+        null=True,
+        blank=True,
+        help_text=(
+            "Ordered recruitment steps. "
+            "Each item: {'entity_id': str, 'categorie': str, 'nom': str}"
+        ),
+    )
+
+    class Meta:
+        db_table = "organisme"
+        verbose_name = "Organisme"
+        verbose_name_plural = "Organismes"
+
+    def __str__(self) -> str:
+        return str(self.id)
