+++ src/web/tests/utils/seed_recruteur_datas.py
@@ -7,7 +7,7 @@ from referentiel.value_objects.verse import Verse
 from domain.candidate.value_objects.statut_candidature import StatutCandidature
 from domain.identite.value_objects.siret import SIRET
 from infrastructure.django_apps.candidate.models.candidature import CandidatureModel
-from infrastructure.django_apps.recruteur.models import OrganismeModel
+from infrastructure.django_apps.recruteur.models.organisme import OrganismeModel
 from infrastructure.django_apps.referentiel.models.metier import MetierModel
 from infrastructure.django_apps.referentiel.models.offer import OfferModel
 from infrastructure.django_apps.users.models import (
